+++ tests/partition_test/test_gen_partition.py
@@ -201,6 +201,27 @@ class TestGenPartition(TestPartitionFixture):
                 / len(list(self._gen_partition_full()))
         self.assertLessEqual(r, 0.15)
 
+    def test_guaranteed(self):
+        ''' Guaranteed. '''
+        optkey = 'NOBATP'
+
+        for wlkey in self.layers:
+            if wlkey.startswith('SSM'):
+                print wlkey
+                self.assertEqual(len(list(self._gen_partition(wlkey=wlkey,
+                                                              optkey=optkey))),
+                                 0)
+
+                part_list = list(self._gen_partition(wlkey=wlkey,
+                                                     optkey=optkey,
+                                                     guaranteed=True))
+                self.assertEqual(len(part_list), 1)
+                part = part_list[0]
+                self.assertEqual(part.size(pe.BATP), 1)
+                self.assertEqual(part.size(pe.INPP), 1)
+                self.assertTrue(part.size(pe.OUTP) == 1
+                                or part.size(pe.OFMP) == 1)
+
     def _part_index_to_coord(self, part):
         ''' Get the mapping from partition index to coordinate. '''
         mapping = {}